add API with hapi
* currently only the route GET on /api/page
* simplifies signaling server
