+++ package.json
@@ -18,8 +18,9 @@
   },
   "homepage": "https://github.com/dunlin/signaling-server",
   "dependencies": {
-    "lodash": "^3.9.3",
-    "node-uuid": "^1.4.2",
+    "hapi": "8.6.1",
+    "lodash": "3.9.3",
+    "moniker": "0.1.2",
     "socket.io": "0.9.16"
   }
 }
